requirements.txt
appdirs==1.4.3
astroid==2.4.2
autopep8==1.5.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
conda==4.3.16
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2
django-allauth==0.42.0
django-ckeditor==6.0.0
django-cleanup==5.0.0
django-cors-headers==3.5.0
django-etc==1.2.0
django-filebrowser==3.13.1
django-grappelli==2.14.2
django-hitcount==1.3.3
django-jalali-date==0.3.1
django-js-asset==1.2.2
django-remote-forms==0.0.1
django-rest-auth==0.9.5
django-tinymce==3.0.2
djangorestframework==3.11.1
flake8==3.8.3
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
jdatetime==3.6.2
jedi==0.17.2
kavenegar==1.1.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
parso==0.7.1
pep517==0.8.2
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycosat==0.6.3
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.5.3
pyparsing==2.4.6
python-decouple==3.3
python-jsonrpc-server==0.4.0
python-language-server==0.35.0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
ujson==3.2.0
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
yapf==0.30.0
